# SwaggerDocumentation.json
{"swagger":"2.0","info":{"description":"Allows to perform Delete, Update, Create and Read customer profile information.","version":"V1","title":"Customer Details","contact":{"name":"Sarabeshwar Chandrasekar","email":"sarabeshwarc@gmail.com"},"license":{}},"host":"localhost:8080","basePath":"/","tags":[{"name":"customer-controller","description":"Set of REST Endpoint for Creating, Retrieving, Updating and Deleting Customer Profiles."}],"paths":{"/Q/customerDetails/createCustomerProfile":{"post":{"tags":["customer-controller"],"summary":"Create Customer new Profile  using  Customer's Unique Account Number and Request JSON.","operationId":"createCustomerProfileUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"accountNumber","in":"header","description":"accountNumber","required":true,"type":"string"},{"in":"body","name":"customerProfileRequst","description":"customerProfileRequst","required":true,"schema":{"$ref":"#/definitions/CustomerProfileRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CreateCustomerProfileResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/Q/customerDetails/deleteCustomerProfile/{profileId}/{accountNumber}":{"delete":{"tags":["customer-controller"],"summary":"Delete a Customer Profile.","operationId":"deleteCustomerProfileUsingDELETE","produces":["application/json"],"parameters":[{"name":"accountNumber","in":"path","description":"accountNumber","required":true,"type":"string"},{"name":"profileId","in":"path","description":"profileId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DeleteCustomerProfile"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/Q/customerDetails/retrieveCustomerprofile/{profileId}/{accountNumber}":{"get":{"tags":["customer-controller"],"summary":"Returns Customer Information based on customer ProfileId and Customer's Unique Account Number","operationId":"retrieveCustomerprofileUsingGET","produces":["application/json"],"parameters":[{"name":"accountNumber","in":"path","description":"accountNumber","required":true,"type":"string"},{"name":"profileId","in":"path","description":"profileId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CustomerInformation"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/Q/customerDetails/updateCustomerProfile":{"put":{"tags":["customer-controller"],"summary":"Update Existing Customer Profile Details.","operationId":"updateCustomerProfileUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"customerAccountNumber","in":"header","description":"customerAccountNumber","required":true,"type":"string"},{"in":"body","name":"customerProfileRequst","description":"customerProfileRequst","required":true,"schema":{"$ref":"#/definitions/CustomerProfileRequest"}},{"name":"profileId","in":"header","description":"profileId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UpdateCustomerProfileResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"CreateCustomerProfileResponse":{"type":"object","required":["customerProfile","status"],"properties":{"customerProfile":{"description":"Persisted Customer Details","$ref":"#/definitions/CustomerProfile"},"status":{"type":"string","example":"Success","description":"Status of the Create Request."}},"title":"CreateCustomerProfileResponse","description":"Class representing create Customer Profile Response"},"CustomerAddress":{"type":"object","properties":{"addressType":{"type":"string"},"postcode":{"type":"integer","format":"int32"},"state":{"type":"string"},"street":{"$ref":"#/definitions/StreetInfo"},"suburb":{"type":"string"},"unitNo":{"type":"integer","format":"int32"}},"title":"CustomerAddress"},"CustomerInformation":{"type":"object","properties":{"address":{"type":"string"},"addressType":{"type":"string"},"customerAccountNumber":{"type":"string"},"dateOfBirth":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"profileId":{"type":"string"}},"title":"CustomerInformation"},"CustomerProfile":{"type":"object","properties":{"accountNumber":{"type":"string"},"changeDate":{"type":"string"},"createdDate":{"type":"string"},"profileId":{"type":"integer","format":"int32"},"status":{"type":"string"},"tenure":{"type":"integer","format":"int32"}},"title":"CustomerProfile"},"CustomerProfileRequest":{"type":"object","required":["address","customerAccountNumber","dataOfBirth","firstName","lastName"],"properties":{"address":{"description":"Contain detailed information about the complete address of the customer along with address type.","$ref":"#/definitions/CustomerAddress"},"customerAccountNumber":{"type":"string","example":"QANT756123S","description":"Represents the Customer's Unique Account Number"},"dataOfBirth":{"type":"string","example":"25-12-1988","description":"Date of Birth of the person"},"firstName":{"type":"string","example":"Sarabeshwar","description":"First name of the person."},"lastName":{"type":"string","example":"Chandrasekar","description":"Last name of the person."}},"title":"CustomerProfileRequest","description":"Class representing to create a new customer profile."},"DeleteCustomerProfile":{"type":"object","required":["createdDate","customerAccountNumber","inactiveDate","profileId","profileStatus","statusMessage","tenure"],"properties":{"createdDate":{"type":"string","example":"23-10-2019","description":"Represents the Customer's profile create date"},"customerAccountNumber":{"type":"string","example":"QANT756123S","description":"Represents the Customer's Unique Account Number"},"inactiveDate":{"type":"string","example":"25-10-2023","description":"Represents the Customer's profile inactive date"},"profileId":{"type":"string","example":123,"description":"Represents the Customer's Unique profileId"},"profileStatus":{"type":"string","example":"QANT756123S","description":"Represents the Customer's profile status like enrolled, eligible, active, inactive"},"statusMessage":{"description":"Represents the status of the call","$ref":"#/definitions/StatusMessage"},"tenure":{"type":"integer","format":"int32","example":5,"description":"Represents the Customer's tenure with Qantas"}},"title":"DeleteCustomerProfile","description":"Class representing to Delete a customer profile."},"StatusMessage":{"type":"object","required":["statusCode","statusDescription"],"properties":{"statusCode":{"type":"integer","format":"int32","example":200,"description":"Represents the status code from the CRM"},"statusDescription":{"type":"string","example":"Successfully got the response from CRM","description":"Represents the statusDescription from the CRM"}},"title":"StatusMessage","description":"Class representing Response Status Message"},"StreetInfo":{"type":"object","properties":{"streetName":{"type":"string"},"streetNo":{"type":"integer","format":"int32"},"streetType":{"type":"string"}},"title":"StreetInfo"},"UpdateCustomerProfileResponse":{"type":"object","properties":{"customerAccountNumber":{"type":"string"},"profileId":{"type":"string"},"statusMessage":{"$ref":"#/definitions/StatusMessage"}},"title":"UpdateCustomerProfileResponse"}}}